Comformer/pyproject.toml
[project]
name = "comformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.19.5, <2",
    "scipy>=1.6.1",
    "jarvis-tools>=2021.07.19",
    "scikit-learn>=0.22.2",
    "matplotlib>=3.4.1",
    "tqdm>=4.60.0",
    "pandas>=1.2.3",
    "pytorch-ignite>=0.4.7",
    "pydantic>=1.8.1,<2",
    "pyparsing>=2.2.1,<3",
    "pandarallel>=1.6.5",
    "e3nn>=0.5.4",
    "nvitop>=1.4.2",
    "pymatgen>=2025.1.24",
    "ipywidgets>=8.1.5",
    "pyarrow>=19.0.0",
    "ipykernel>=6.29.5",
    "structlog>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "flake8>=3.9.1",
    "pycodestyle>=2.7.0",
    "pydocstyle>=6.0.0",
]
